Jenkins Task-2_Sridhar Jagadeesan
Techstacks needs to be used : 

●	AWS EC2
●	Github
●	Jenkins

Create a simple script file and push it to repo. Create a project in Jenkins connected to your GitHub repository. When a commit is made to your repo, automatically build must get triggered from Jenkins and the output must be shared to me via email.